fix: move create_retriever to retrieval_chain
