+++ aidial_rag/retrieval_chain.py
@@ -1,12 +1,16 @@
+import asyncio
 from dataclasses import dataclass
 from itertools import groupby
 from operator import itemgetter
-from typing import Any, Dict, List, Set
+from typing import Any, Callable, Dict, List, Set
 
+from langchain.retrievers import EnsembleRetriever
 from langchain.schema import BaseRetriever, Document
 from langchain.schema.runnable import RunnablePassthrough
+from langchain_core.retrievers import RetrieverLike
 from langchain_core.runnables import Runnable, chain
 
+from aidial_rag.dial_config import DialConfig
 from aidial_rag.document_record import DocumentRecord
 from aidial_rag.image_processor.base64 import pil_image_as_base64
 from aidial_rag.image_processor.extract_pages import (
@@ -14,8 +18,16 @@ from aidial_rag.image_processor.extract_pages import (
     extract_pages_gen,
 )
 from aidial_rag.index_record import ChunkMetadata, RetrievalType
+from aidial_rag.indexing_config import IndexingConfig
 from aidial_rag.qa_chain_config import ChatChainConfig
 from aidial_rag.retrieval_api import RetrievalResults
+from aidial_rag.retrievers.all_documents_retriever import AllDocumentsRetriever
+from aidial_rag.retrievers.bm25_retriever import BM25Retriever
+from aidial_rag.retrievers.description_retriever.description_retriever import (
+    DescriptionRetriever,
+)
+from aidial_rag.retrievers.multimodal_retriever import MultimodalRetriever
+from aidial_rag.retrievers.semantic_retriever import SemanticRetriever
 
 
 @dataclass(frozen=True, order=True)
@@ -156,10 +168,93 @@ async def create_retrieval_results(
     )
 
 
+def _make_retrieval_stage_default(
+    retriever: BaseRetriever, stage_name: str
+) -> BaseRetriever:
+    """Default stage maker that returns the retriever as is."""
+    return retriever
+
+
+def create_retriever(
+    dial_config: DialConfig,
+    document_records: List[DocumentRecord],
+    indexing_config: IndexingConfig,
+    make_retrieval_stage: Callable[
+        [BaseRetriever, str], BaseRetriever
+    ] = _make_retrieval_stage_default,
+) -> BaseRetriever:
+    if not AllDocumentsRetriever.is_within_limit(document_records):
+        semantic_retriever = make_retrieval_stage(
+            SemanticRetriever.from_doc_records(document_records, 7),
+            "Embeddings search",
+        )
+        retrievers: List[RetrieverLike] = [semantic_retriever]
+        weights = [1.0]
+
+        if BM25Retriever.has_index(document_records):
+            bm25_retriever = make_retrieval_stage(
+                BM25Retriever.from_doc_records(document_records, 7),
+                "Keywords search",
+            )
+            retrievers.append(bm25_retriever)
+            weights.append(1.0)
+
+        if MultimodalRetriever.has_index(document_records):
+            assert indexing_config.multimodal_index
+            multimodal_retriever = make_retrieval_stage(
+                MultimodalRetriever.from_doc_records(
+                    dial_config,
+                    indexing_config.multimodal_index,
+                    document_records,
+                    7,
+                ),
+                "Multimodal search",
+            )
+            retrievers.append(multimodal_retriever)
+            weights.append(1.0)
+
+        if DescriptionRetriever.has_index(document_records):
+            description_retriever = make_retrieval_stage(
+                DescriptionRetriever.from_doc_records(document_records, 7),
+                "Page image search",
+            )
+            retrievers.append(description_retriever)
+            weights.append(1.0)
+
+        retriever = make_retrieval_stage(
+            EnsembleRetriever(
+                retrievers=retrievers,
+                weights=weights,
+            ),
+            "Combined search",
+        )
+    else:
+        retriever = make_retrieval_stage(
+            AllDocumentsRetriever.from_doc_records(document_records),
+            "All documents",
+        )
+
+    return retriever
+
+
 async def create_retrieval_chain(
+    dial_config: DialConfig,
+    indexing_config: IndexingConfig,
+    document_records: List[DocumentRecord],
     query_chain: Runnable[Dict[str, Any], str],
-    retriever: BaseRetriever,
+    make_retrieval_stage: Callable[
+        [BaseRetriever, str], BaseRetriever
+    ] = _make_retrieval_stage_default,
 ) -> Runnable[Dict[str, Any], Dict[str, Any]]:
+    retriever = await asyncio.get_running_loop().run_in_executor(
+        None,
+        create_retriever,
+        dial_config,
+        document_records,
+        indexing_config,
+        make_retrieval_stage,
+    )
+
     retrieval_chain = (
         RunnablePassthrough()
         .assign(query=query_chain)
